code enhancement
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C36680

+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Entities/SupplierZone/SupplierZoneServiceQuery.cs
@@ -7,7 +7,7 @@ namespace TOne.WhS.BusinessEntity.Entities
     public class SupplierZoneServiceQuery
     {
        
-        public List<int> ZoneIds { get; set; }
+        public List<long> ZoneIds { get; set; }
         public int SupplierId { get; set; }
         public DateTime EffectiveOn { get; set; }
